:recycle: Move report tests to new file
